Card Game Basic Implementation
When 'E' is pressed, UI will appear for a card game. Pressing the card that has the higher value will prompt 'You Win!' text or 'You lose" text if you select the card with a lower value. Game can be replayed as much as the player likes after making a selection, or return to normal gameplay. Visuals still need to be added.