Interaction Text Update 2
Yes this is a no-descriptive name, there were really only a few small changes. Interaction text now displays two boxes of text but cannot be more than that and still runs on a timer. The reason for committing this is that attempting to implement a while loop to check if all text had been gone through either A) wouldn't print the text to the screen or B) crash the game due to an infinite loop.

Next commit will try to have the text advance via player input with an event dispatcher, rather than a while loop