Interaction Text functionality complete
Text now waits until player presses E again to advance text, or if the whole array of text has been looped through, either exit the interaction and return to normal gameplay, or start a minigame.
Next updates will be creating a locked door that the player has to beat the minigame to get through, entering the win state once the play meets the right conditions, and (possibly) giving text once the player exits the minigame, depending on whether they won or lost.